pyproject.toml
[tool.poetry]
name = "totoro"
version = "0.1.0"
description = "Studio Ghibli API wrapper"
authors = ["Abhishek Juneja <abhishek.juneja145@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
django = "2.2"
django-configurations = "^2.2"
django-extensions = "^2.2.9"
ipython = "^7.14.0"
bpython = "^0.19"
djangorestframework = "^3.11.0"
django-redis = "^4.11.0"

[tool.poetry.dev-dependencies]
pre-commit = "^2.3.0"
django-debug-toolbar = "^2.2"
pytest-pythonpath = "^0.7.3"
pytest-django = "^3.9.0"
werkzeug = "^1.0.1"
drf-yasg = {extras = ["validation"], version = "^1.17.1"}

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
